Swtiched ordering of step 2 and step 3 (name/email before major/intro)
