+++ packages/client/src/pages/SignUp/wizards/ParentWizard/forms/ParentStep2.js
@@ -1,14 +1,10 @@
-import React, { useState } from 'react';
+import React from 'react';
 import styled from 'styled-components';
-
-import Select from '@material-ui/core/Select';
 import MenuItem from '@material-ui/core/MenuItem';
 import InputLabel from '@material-ui/core/InputLabel';
 import TextField from '@material-ui/core/TextField';
-import Checkbox from '@material-ui/core/Checkbox';
-import ListItemText from '@material-ui/core/ListItemText';
-
-import { tags, subjects } from '../../../../../constants.js';
+import Select from '@material-ui/core/Select';
+import { timeZones } from '../../../../../constants.js';
 
 const WizardInput = styled.div`
     margin-bottom: 1em;
@@ -27,32 +23,9 @@ const SelectMenuProps = {
 
 const ParentStep2 = (props) => {
 
-    const [state, setState] = useState({
-        selectedSubjects: [],
+    const timeZoneMenuItems = timeZones.map(item => {
+        return <MenuItem key={item.value} value={item.timezone}>{item.timezone}</MenuItem>;
     });
-
-    const { selectedSubjects } = state;
-
-    const gradeLevelMenuItems = tags.map((item, index) => (
-        <MenuItem key={index}>
-            <ListItemText primary={item.label} />
-        </MenuItem>
-    ));
-
-    const subjectsMenuItems = subjects.map((item, index) => (
-        <MenuItem key={index} value={item.value}>
-            <Checkbox
-                checked={selectedSubjects.indexOf(item.value) > -1}
-                value={item.value}
-            />
-            <ListItemText primary={item.label} />
-        </MenuItem>
-    ));
-
-    const handleChange = (event) => {
-        setState({ ...state, [event.target.name]: event.target.value });
-        props.updateRegisteredChild(props.index, { ...state, [event.target.name]: event.target.value });
-    }
 
     return (
         <div>
@@ -60,45 +33,45 @@ const ParentStep2 = (props) => {
             <WizardInput>
                 <TextField
                     fullWidth
-                    label="Student Name"
-                    name="studentName"
-                    onChange={handleChange}
-                    value={props.data.registeredChildren[props.index].studentName}
+                    label="Parent Name"
+                    name="parentName"
+                    onChange={props.handleChange}
+                    value={props.data.parentName}
                     required
                 />
             </WizardInput>
             <WizardInput>
                 <TextField
                     fullWidth
-                    label="Student Email"
-                    name="studentEmail"
-                    onChange={handleChange}
-                    value={props.data.registeredChildren[props.index].studentEmail}
+                    label="Parent Email"
+                    name="parentEmail"
+                    onChange={props.handleChange}
+                    value={props.data.parentEmail}
                     required
                 />
             </WizardInput>
             <WizardInput>
-                <InputLabel id="wizard-time-zone" required>Grade Level</InputLabel>
-                <Select
-                    children={gradeLevelMenuItems}
+                <TextField
                     fullWidth
-                    name="gradeLevel"
-                    onChange={handleChange}
-                    value={props.data.registeredChildren[props.index].gradeLevel}
+                    label="Parent Phone Number"
+                    name="parentPhoneNumber"
+                    onChange={props.handleChange}
+                    value={props.data.parentPhoneNumber}
                     required
                 />
             </WizardInput>
             <WizardInput>
-                <InputLabel id="wizard-time-zone" required>Student Seeking Assistance In</InputLabel>
+                <InputLabel id="wizard-time-zone" required>Time Zone</InputLabel>
                 <Select
-                    children={subjectsMenuItems}
+                    children={timeZoneMenuItems}
+                    displayEmpty
                     fullWidth
+                    labelId="wizard-preferred-subjects"
                     MenuProps={SelectMenuProps}
-                    multiple
-                    name='selectedSubjects'
-                    onChange={handleChange}
-                    renderValue={(selected) => selected.join(', ')}
-                    value={props.data.registeredChildren[props.index].selectedSubjects}
+                    name="timeZone"
+                    onChange={props.handleChange}
+                    renderValue={(selected) => selected}
+                    value={props.data.timeZone}
                     required
                 />
             </WizardInput>
